libambit: Separate struct declarations from rest of API
Exposing non-opaque structs makes it hard to keep a stable ABI and API,
while adding support for new features and models and device
generations. When the user and library are in the same repository and
always share the same version, there isn't really any issues, but in
order to be able to use libambit as an external library, lets make it
easy to not accidentally use any structs directly.

In later patches, accessor API will be added to get the data previously
read directly from the structs.

The structs should some how be declared not part of any stable API,
while functions, and getters and setters will.
