+++ src/libambit/libambit.h
@@ -26,353 +26,11 @@
 #include <stdbool.h>
 #include <time.h>
 
+#include "libambit-types.h"
+
 #ifdef __cplusplus /* If this is a C++ compiler, use C linkage */
 extern "C" {
 #endif
-
-typedef struct ambit_object_s ambit_object_t;
-
-typedef struct ambit_device_info_s {
-    char *name;                                        /* UTF-8 */
-    char *model;                                       /* UTF-8 */
-    char *serial;                                      /* UTF-8 */
-    uint8_t fw_version[4];
-    uint8_t hw_version[4];
-
-    const char *path;                   /* file system encoding */
-    uint16_t    vendor_id;
-    uint16_t    product_id;
-    bool        is_supported;
-    int         access_status;
-
-    struct ambit_device_info_s *next;
-} ambit_device_info_t;
-
-typedef struct ambit_device_status_s {
-    uint8_t  charge;
-} ambit_device_status_t;
-
-typedef struct ambit_personal_settings_s {
-    uint8_t  sportmode_button_lock;
-    uint8_t  timemode_button_lock;
-    uint16_t compass_declination;
-    float compass_declination_f;      /* Ambit3: 32 bit float containing the radian */
-    uint8_t  units_mode;
-    struct {
-        uint8_t pressure;
-        uint8_t altitude;
-        uint8_t distance;
-        uint8_t height;
-        uint8_t temperature;
-        uint8_t verticalspeed;
-        uint8_t weight;
-        uint8_t compass;
-        uint8_t heartrate;
-        uint8_t speed;
-    } units;
-    uint8_t  gps_position_format;
-    uint8_t  language;
-    uint8_t  navigation_style;
-    uint8_t  sync_time_w_gps;
-    uint8_t  time_format;
-    struct {
-        uint8_t hour;
-        uint8_t minute;
-    } alarm;
-    uint8_t  alarm_enable;
-    struct {
-        uint8_t hour;
-        uint8_t minute;
-    } dual_time;
-    uint8_t  date_format;
-    uint8_t  tones_mode;
-    uint8_t  backlight_mode;
-    uint8_t  backlight_brightness;
-    uint8_t  display_brightness;
-    uint8_t  display_is_negative;
-    uint16_t weight;                  /* kg scale 0.01 */
-    uint16_t birthyear;
-    uint8_t  max_hr;
-    uint8_t  rest_hr;
-    uint8_t  fitness_level;
-    uint8_t  is_male;
-    uint8_t  length;
-    uint8_t  alti_baro_mode;
-    uint8_t  storm_alarm;
-    uint8_t  fused_alti_disabled;
-    uint16_t bikepod_calibration;     /* scale 0.0001 */
-    uint16_t bikepod_calibration2;    /* scale 0.0001 */
-    uint16_t bikepod_calibration3;    /* scale 0.0001 */
-    uint16_t footpod_calibration;     /* scale 0.0001 */
-    uint8_t  automatic_bikepower_calib;
-    uint8_t  automatic_footpod_calib;
-    uint8_t  training_program;
-} ambit_personal_settings_t;
-
-typedef struct ambit_log_date_time_s {
-    uint16_t year;
-    uint8_t  month;
-    uint8_t  day;
-    uint8_t  hour;
-    uint8_t  minute;
-    uint16_t msec;
-} ambit_date_time_t;
-
-typedef enum ambit_log_sample_type_e {
-    ambit_log_sample_type_periodic = 0x0200,
-    ambit_log_sample_type_logpause = 0x0304,
-    ambit_log_sample_type_logrestart = 0x0305,
-    ambit_log_sample_type_ibi = 0x0306,
-    ambit_log_sample_type_ttff = 0x0307,
-    ambit_log_sample_type_distance_source = 0x0308,
-    ambit_log_sample_type_lapinfo = 0x0309,
-    ambit_log_sample_type_altitude_source = 0x030d,
-    ambit_log_sample_type_gps_base = 0x030f,
-    ambit_log_sample_type_gps_small = 0x0310,
-    ambit_log_sample_type_gps_tiny = 0x0311,
-    ambit_log_sample_type_time = 0x0312,
-    ambit_log_sample_type_swimming_turn = 0x0314,
-    ambit_log_sample_type_swimming_stroke = 0x0315,
-    ambit_log_sample_type_activity = 0x0318,
-    ambit_log_sample_type_cadence_source = 0x031a,
-    ambit_log_sample_type_position = 0x031b,
-    ambit_log_sample_type_fwinfo = 0x031c,
-    ambit_log_sample_type_unknown = 0xf000
-} ambit_log_sample_type_t;
-
-typedef enum ambit_log_sample_periodic_type_e {
-    ambit_log_sample_periodic_type_latitude = 0x01,
-    ambit_log_sample_periodic_type_longitude = 0x02,
-    ambit_log_sample_periodic_type_distance = 0x03,
-    ambit_log_sample_periodic_type_speed = 0x04,
-    ambit_log_sample_periodic_type_hr = 0x05,
-    ambit_log_sample_periodic_type_time = 0x06,
-    ambit_log_sample_periodic_type_gpsspeed = 0x07,
-    ambit_log_sample_periodic_type_wristaccspeed = 0x08,
-    ambit_log_sample_periodic_type_bikepodspeed = 0x09,
-    ambit_log_sample_periodic_type_ehpe = 0x0a,
-    ambit_log_sample_periodic_type_evpe = 0x0b,
-    ambit_log_sample_periodic_type_altitude = 0x0c,
-    ambit_log_sample_periodic_type_abspressure = 0x0d,
-    ambit_log_sample_periodic_type_energy = 0x0e,
-    ambit_log_sample_periodic_type_temperature = 0x0f,
-    ambit_log_sample_periodic_type_charge = 0x10,
-    ambit_log_sample_periodic_type_gpsaltitude = 0x11,
-    ambit_log_sample_periodic_type_gpsheading = 0x12,
-    ambit_log_sample_periodic_type_gpshdop = 0x13,
-    ambit_log_sample_periodic_type_gpsvdop = 0x14,
-    ambit_log_sample_periodic_type_wristcadence = 0x15,
-    ambit_log_sample_periodic_type_snr = 0x16,
-    ambit_log_sample_periodic_type_noofsatellites = 0x17,
-    ambit_log_sample_periodic_type_sealevelpressure = 0x18,
-    ambit_log_sample_periodic_type_verticalspeed = 0x19,
-    ambit_log_sample_periodic_type_cadence = 0x1a,
-    ambit_log_sample_periodic_type_bikepower = 0x1f,
-    ambit_log_sample_periodic_type_swimingstrokecnt = 0x20,
-    ambit_log_sample_periodic_type_ruleoutput1 = 0x64,
-    ambit_log_sample_periodic_type_ruleoutput2 = 0x65,
-    ambit_log_sample_periodic_type_ruleoutput3 = 0x66,
-    ambit_log_sample_periodic_type_ruleoutput4 = 0x67,
-    ambit_log_sample_periodic_type_ruleoutput5 = 0x68
-} ambit_log_sample_periodic_type_t;
-
-typedef struct ambit_log_sample_periodic_value_s {
-    ambit_log_sample_periodic_type_t type;
-    union {
-        int32_t  latitude;          /* degree, scale: 0.0000001, -90 <= latitude <= 90 */
-        int32_t  longitude;         /* degree, scale: 0.0000001, -180 <= latitude <= 180 */
-        uint32_t distance;          /* meter, scale: 1 */
-        uint16_t speed;             /* m/s scale: 0.01, ignore if 0xffff */
-        uint8_t  hr;                /* bpm, ignore if 0xff */
-        uint32_t time;              /* msec */
-        uint16_t gpsspeed;          /* m/s scale: 0.01 */
-        uint16_t wristaccspeed;     /* m/s scale: 0.01 */
-        uint16_t bikepodspeed;      /* m/s scale: 0.01 */
-        uint32_t ehpe;              /* meters scale: 0.01 */
-        uint32_t evpe;              /* meters scale: 0.01 */
-        int16_t  altitude;          /* meters, -1000 <= altitude <= 10000 */
-        uint16_t abspressure;       /* hpa scale 0.1 */
-        uint16_t energy;            /* hcal/min */
-        int16_t  temperature;       /* celsius scale 0.1, -100 <= temp <= 100 */
-        uint8_t  charge;            /* percent */
-        int32_t  gpsaltitude;       /* meters scale: 0.01, -1000 <= altitude <= 10000 */
-        uint16_t gpsheading;        /* degrees, scale: 0.01, 0 <= heading <= 360 */
-        uint8_t  gpshdop;           /* ? */
-        uint8_t  gpsvdop;           /* ? */
-        uint16_t wristcadence;      /* rpm, ignore if 0xffff */
-        uint8_t  snr[16];           /* ? */
-        uint8_t  noofsatellites;    /* ignore if 0xff */
-        int16_t  sealevelpressure;  /* hpa scale: 0.1 */
-        int16_t  verticalspeed;     /* m/s scale: 0.01 */
-        uint8_t  cadence;           /* rpm, ignore if 0xff */
-        uint16_t bikepower;         /* watt */
-        uint32_t swimingstrokecnt;  /* counter */
-        int32_t  ruleoutput1;       /* ignore if 0x80000000 */
-        int32_t  ruleoutput2;       /* ignore if 0x80000000 */
-        int32_t  ruleoutput3;       /* ignore if 0x80000000 */
-        int32_t  ruleoutput4;       /* ignore if 0x80000000 */
-        int32_t  ruleoutput5;       /* ignore if 0x80000000 */
-    } u;
-} ambit_log_sample_periodic_value_t;
-
-typedef struct ambit_log_gps_satellite_s {
-    uint8_t sv;
-    uint8_t snr;
-    uint8_t state;
-} ambit_log_gps_satellite_t;
-
-typedef struct ambit_log_sample_s {
-    ambit_log_sample_type_t type;  /* type of sample */
-    uint32_t time;                 /* time from zero */
-    ambit_date_time_t utc_time;
-    union {
-        struct {
-            uint8_t value_count;
-            ambit_log_sample_periodic_value_t *values;
-        } periodic;
-        struct {
-            uint8_t ibi_count;
-            uint16_t ibi[32];
-        } ibi;
-        uint16_t ttff;
-        uint8_t  distance_source;               /* 0x00 = Bikepod,
-                                                   0x01 = Footpod,
-                                                   0x02 = GPS,
-                                                   0x03 = Wrist,
-                                                   0x04 = Indoorswimming,
-                                                   0x05 = Outdoorswimming */
-        struct {
-            uint8_t event_type;                 /* 0x01 = manual lap,
-                                                   0x14 = high interval end,
-                                                   0x15 = low interval end,
-                                                   0x16 = interval start,
-                                                   0x1e = pause,
-                                                   0x1f = start */
-            ambit_date_time_t date_time;
-            uint32_t duration;                  /* ms */
-            uint32_t distance;                  /* meters */
-        } lapinfo;
-        struct {
-            uint8_t source_type;                /* 0x04 = pressure */
-            int16_t altitude_offset;
-            int16_t pressure_offset;
-        } altitude_source;
-        struct {
-            uint16_t navvalid;
-            uint16_t navtype;
-            ambit_date_time_t utc_base_time;
-            int32_t  latitude;                  /* degree, scale: 0.0000001, -90 <= latitude <= 90 */
-            int32_t  longitude;                 /* degree, scale: 0.0000001, -180 <= latitude <= 180 */
-            int32_t  altitude;                  /* meters scale: 0.01, -1000 <= altitude <= 10000 */
-            uint16_t speed;                     /* m/s scale: 0.01 */
-            uint16_t heading;                   /* degrees, scale: 0.01, 0 <= heading <= 360 */
-            uint32_t ehpe;                      /* meters scale: 0.01 */
-            uint8_t  noofsatellites;            /* */
-            uint8_t  hdop;                      /* ? scale: 0.2 */
-            uint8_t  satellites_count;
-            ambit_log_gps_satellite_t *satellites;
-        } gps_base;
-        struct {
-            uint8_t  noofsatellites;            /* */
-            int32_t  latitude;                  /* degree, scale: 0.0000001, -90 <= latitude <= 90 */
-            int32_t  longitude;                 /* degree, scale: 0.0000001, -180 <= latitude <= 180 */
-            uint32_t ehpe;                      /* meters scale: 0.01 */
-        } gps_small;
-        struct {
-            int32_t  latitude;                  /* degree, scale: 0.0000001, -90 <= latitude <= 90 */
-            int32_t  longitude;                 /* degree, scale: 0.0000001, -180 <= latitude <= 180 */
-            uint32_t ehpe;                      /* meters scale: 0.01 */
-            uint8_t  unknown;                   /* Unknown data */
-        } gps_tiny;
-        struct {
-            uint8_t  hour;
-            uint8_t  minute;
-            uint8_t  second;
-        } time;
-        struct {
-            uint32_t distance;                  /* Total distance, meters scale: 0.01 */
-            uint16_t lengths;                   /* Total pool lengths */
-            uint16_t classification[4];
-            uint8_t  style;                     /* (style of previous length)
-                                                   0x00 = Other,
-                                                   0x01 = Butterfly,
-                                                   0x02 = Backstroke,
-                                                   0x03 = Breaststroke,
-                                                   0x04 = Freestyle,
-                                                   0x05 = Drill */
-        } swimming_turn;
-        struct {
-            uint16_t activitytype;
-            uint32_t custommode;
-        } activity;
-        uint8_t cadence_source;                 /* 0x40 = Wrist */
-        struct {
-            int32_t  latitude;                  /* degree, scale: 0.0000001, -90 <= latitude <= 90 */
-            int32_t  longitude;                 /* degree, scale: 0.0000001, -180 <= latitude <= 180 */
-        } position;
-        struct {
-            uint8_t version[4];
-            ambit_date_time_t build_date;
-        } fwinfo;
-        struct {
-            size_t datalen;
-            uint8_t *data;
-        } unknown;
-    } u;
-} ambit_log_sample_t;
-
-typedef struct ambit_log_header_s {
-    ambit_date_time_t date_time;
-    uint32_t duration;              /* ms */
-    uint16_t ascent;                /* m */
-    uint16_t descent;               /* m */
-    uint32_t ascent_time;           /* ms */
-    uint32_t descent_time;          /* ms */
-    uint32_t recovery_time;         /* ms */
-    uint16_t speed_avg;             /* m/h */
-    uint16_t speed_max;             /* m/h */
-    uint32_t speed_max_time;        /* ms */
-    int16_t  altitude_max;          /* m */
-    int16_t  altitude_min;          /* m */
-    uint32_t altitude_max_time;     /* ms */
-    uint32_t altitude_min_time;     /* ms */
-    uint8_t  heartrate_avg;         /* bpm */
-    uint8_t  heartrate_max;         /* bpm */
-    uint8_t  heartrate_min;         /* bpm */
-    uint32_t heartrate_max_time;    /* ms */
-    uint32_t heartrate_min_time;    /* ms */
-    uint8_t  peak_training_effect;  /* effect scale 0.1 */
-    uint8_t  activity_type;
-    char    *activity_name;         /* name of activity in UTF-8 */
-    int16_t  temperature_max;       /* degree celsius scale 0.1 */
-    int16_t  temperature_min;       /* degree celsius scale 0.1 */
-    uint32_t temperature_max_time;  /* ms */
-    uint32_t temperature_min_time;  /* ms */
-    uint32_t distance;              /* m */
-    uint32_t samples_count;         /* number of samples in log */
-    uint16_t energy_consumption;    /* kcal */
-    uint32_t first_fix_time;        /* ms */
-    uint8_t  battery_start;         /* percent */
-    uint8_t  battery_end;           /* percent */
-    uint32_t distance_before_calib; /* m */
-
-    uint8_t  unknown1[5];
-    uint8_t  unknown2;
-    uint8_t  cadence_max;           /* rpm */
-    uint8_t  cadence_avg;           /* rpm */
-    uint8_t  unknown3[2];
-    uint16_t swimming_pool_lengths;
-    uint32_t cadence_max_time;      /* ms */
-    uint32_t swimming_pool_length;  /* m */
-    uint8_t  unknown5[4];
-    uint8_t  unknown6[24];
-} ambit_log_header_t;
-
-typedef struct ambit_log_entry_s {
-    ambit_log_header_t header;
-    uint32_t samples_count;
-    ambit_log_sample_t *samples;
-} ambit_log_entry_t;
 
 /** \brief Create a list of all known Ambit clocks on the system
  *
